+++ tests/motion_test.py
@@ -16,7 +16,6 @@
 import numpy as np
 import pytest
 from numpy.testing import assert_allclose, assert_equal
-
 from pystrata import motion
 
 from . import FPATH_DATA
@@ -76,8 +75,8 @@ def test_ts_fft_with_tf(tsm):
 def test_ts_fa_normalize(tsm):
     """Test the normalization of the Fourier amplitudes with Parsevla's theorem."""
     assert_allclose(
-        np.trapz(tsm.accels**2, dx=tsm.time_step),
-        2 * np.trapz(np.abs(tsm.fourier_amps) ** 2, x=tsm.freqs),
+        np.trapezoid(tsm.accels**2, dx=tsm.time_step),
+        2 * np.trapezoid(np.abs(tsm.fourier_amps) ** 2, x=tsm.freqs),
     )