+++ tests/notebooks_test.py
@@ -36,7 +36,7 @@ def test_notebook(fpath):
     with fpath.open() as fp:
         nb = nbformat.read(fp, as_version=4)
 
-    ep = ExecutePreprocessor(timeout=600, kernel_name="python3")
+    ep = ExecutePreprocessor(timeout=600, kernel_name="py312")
 
     try:
         ep.preprocess(nb, {"metadata": {"path": fpath.parent}})